+++ app/views/layouts/_dropdown_menu.html.slim
@@ -1,0 +1,4 @@
+div.border-2.flex.justify-between.px-5
+  = image_tag  "LSGCO_logo.png", class:"h-32 text-center"
+  button 
+    =image_tag src="icons/cancel.svg", size: "30px"
