+++ app/views/users/profile.html.slim
@@ -1,13 +1,13 @@
-.flex.flex-col.justify-start.p-10.w-4/5
-  .flex.w-full.h-full
-    .flex.flex-col.w-2/5.h-full
-      div.h-1/3.overflow-y-auto
+.flex.flex-col.justify-start.p-10.w-full.lg:w-4/5.lg:h-full
+  .flex.flex-col-reverse.w-full.gap-10.lg:flex-row.lg:h-full
+    .flex.flex-col.w-full.lg:w-2/5.lg:h-full
+      div.overflow-y-auto.lg:h-1/3
         h1.title-normal.text-center
           |Tava biedra naudas bilance
         .border-2.rounded-xl.p-5.w-full
           =render "bilance_data"
         
-      div.h-2/3.overflow-y-auto.mt-4
+      div.h-2/3.overflow-y-auto.mt-10
         h1.title-normal.text-center
           |Tuvākie pasākumi
         .border-2.rounded-xl.p-5.w-full.divide-y-2
@@ -21,10 +21,10 @@
             tbody
             -@events.each do |event|
               =render partial: "events/event_list_item", locals: {event: event}
-    .flex.flex-col.h-full.w-3/5.pl-10 
+    .flex.flex-col.h-full.lg:w-3/5.lg:pl-10 
       .flex.justify-between.items-center
         h1.title-normal 
           |Par tevi
         =link_to "Labot datus" ,edit_user_path(@user.id), class:"btn-ternary"
-      .border-2.rounded-xl.h-full.p-5.w-full.text-end
+      .border-2.rounded-xl.h-60.p-5.w-full.text-end.overflow-auto.lg:h-full
         =render "user_data"
